More cleanup in `esp-radio`
