+++ esp-radio/src/lib.rs
@@ -175,7 +175,6 @@ unstable_module! {
     #[cfg(feature = "ieee802154")]
     pub mod ieee802154;
 }
-pub mod config;
 
 pub(crate) mod common_adapter;
 
@@ -203,7 +202,33 @@ const _: () = {
     };
 };
 
-pub(crate) const CONFIG: config::Config = config::Config {
+#[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+/// Tunable parameters for the Wi-Fi driver
+#[allow(unused)] // currently there are no ble tunables
+pub(crate) struct Config {
+    pub(crate) rx_queue_size: usize,
+    pub(crate) tx_queue_size: usize,
+    pub(crate) static_rx_buf_num: usize,
+    pub(crate) dynamic_rx_buf_num: usize,
+    pub(crate) static_tx_buf_num: usize,
+    pub(crate) dynamic_tx_buf_num: usize,
+    pub(crate) ampdu_rx_enable: bool,
+    pub(crate) ampdu_tx_enable: bool,
+    pub(crate) amsdu_tx_enable: bool,
+    pub(crate) rx_ba_win: usize,
+    pub(crate) max_burst_size: usize,
+    pub(crate) country_code: &'static str,
+    pub(crate) country_code_operating_class: u8,
+    pub(crate) mtu: usize,
+    pub(crate) listen_interval: u16,
+    pub(crate) beacon_timeout: u16,
+    pub(crate) ap_beacon_timeout: u16,
+    pub(crate) failure_retry_cnt: u8,
+    pub(crate) scan_method: u32,
+}
+
+pub(crate) const CONFIG: Config = Config {
     rx_queue_size: esp_config_int!(usize, "ESP_RADIO_CONFIG_RX_QUEUE_SIZE"),
     tx_queue_size: esp_config_int!(usize, "ESP_RADIO_CONFIG_TX_QUEUE_SIZE"),
     static_rx_buf_num: esp_config_int!(usize, "ESP_RADIO_CONFIG_STATIC_RX_BUF_NUM"),
